# back/api/openapi.yaml
openapi: 3.0.3
info:
  title: Plant Management API
  description: API REST pour la gestion de plantes domestiques
  version: 1.0.0
  contact:
    name: Plant Team
servers:
  - url: http://localhost:8080/api/v1
    description: Development (Quarkus local)
  # a remplacer par l'URL du cluster Kubernetes université ?
  - url: http://plant-backend.etu-k8s.lan/api/v1
    description: Production (Kubernetes cluster)
...
security:
  - bearerAuth: []
paths:
...
  /plants:
    get:
      tags: [Plant]
      summary: Lister mes plantes
      parameters:
        - name: roomId
          in: query
          schema:
            type: string
            format: uuid
          description: Filtrer par pièce
      responses:
        '200':
          description: Liste des plantes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlantResponse'
    post:
      tags: [Plant]
      summary: Ajouter une plante
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlantRequest'
      responses:
        '201':
          description: Plante ajoutée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantResponse'
  /plants/{id}:
    get:
      tags: [Plant]
      summary: Détails d'une plante
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Détails de la plante
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantDetailResponse'
    patch:
      tags: [Plant]
      summary: Modifier une plante
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlantRequest'
      responses:
        '200':
          description: Plante modifiée
    delete:
      tags: [Plant]
      summary: Supprimer une plante
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Plante supprimée
  /plants/{id}/photo:
    post:
      tags: [Plant]
      summary: Uploader une photo
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Photo uploadée
          content:
            application/json:
              schema:
                type: object
                properties:
                  photoUrl:
                    type: string
  /plants/{id}/water:
    post:
      tags: [Plant]
      summary: Marquer comme arrosée
      description: Raccourci pour créer un CareLog WATERING
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Arrosage enregistré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantResponse'

  # ==================== CARE LOGS ====================
  /plants/{plantId}/care-logs:
    get:
      tags: [CareLog]
      summary: Historique des soins
      parameters:
        - name: plantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Liste des soins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CareLogResponse'
    post:
      tags: [CareLog]
      summary: Enregistrer un soin
      parameters:
        - name: plantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCareLogRequest'
      responses:
        '201':
          description: Soin enregistré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CareLogResponse'

  # ==================== NOTIFICATIONS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    mcpApiKey:
      type: apiKey
      in: header
      name: X-MCP-API-Key
      description: API Key pour l'accès MCP (Goose/Mistral)
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    RoomResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
        plantCount:
          type: integer
...
    SpeciesResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        trefleId:
          type: integer
        commonName:
          type: string
        scientificName:
          type: string
        family:
          type: string
        imageUrl:
          type: string
...
    CreatePlantRequest:
      type: object
      required: [speciesId]
      properties:
        speciesId:
          type: string
          format: uuid
        roomId:
          type: string
          format: uuid
        nickname:
          type: string
          maxLength: 100
        wateringIntervalDays:
          type: integer
          minimum: 1
          maximum: 365
          default: 7
    UpdatePlantRequest:
      type: object
      properties:
        roomId:
          type: string
          format: uuid
          nullable: true
        nickname:
          type: string
        wateringIntervalDays:
          type: integer
        notes:
          type: string
    PlantResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nickname:
          type: string
        species:
          $ref: '#/components/schemas/SpeciesResponse'
        room:
          $ref: '#/components/schemas/RoomResponse'
        photoUrl:
          type: string
        lastWatered:
          type: string
          format: date-time
        nextWatering:
          type: string
          format: date-time
        needsWater:
          type: boolean
        acquiredAt:
          type: string
          format: date
    PlantDetailResponse:
      allOf:
        - $ref: '#/components/schemas/PlantResponse'
        - type: object
          properties:
            wateringIntervalDays:
              type: integer
            notes:
              type: string
            createdAt:
              type: string
              format: date-time
            recentCareLogs:
              type: array
              items:
                $ref: '#/components/schemas/CareLogResponse'

    # ===== CARE LOG =====
    CreateCareLogRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [WATERING, FERTILIZING, REPOTTING, PRUNING, TREATMENT, NOTE]
        notes:
          type: string
    CareLogResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        action:
          type: string
        notes:
          type: string
        performedBy:
          type: object
          properties:
            id:
              type: string
              format: uuid
            displayName:
              type: string
        performedAt:
          type: string
          format: date-time

    # ===== NOTIFICATION =====
...